Modularized code to messaging & ancillary_methods
